IBAN validation: Use additional Regex from schema
The SEPA::IBANValidator uses the iban-tools gem. This third-party gem
is fine, but a bit sloppy: It strips all spaces and upcases the string
before doing further checks. As a result, it allows something like
"de 87  20 05 0000 12 34 5678 90  " to be valid IBAN. Because PAIN is
more strict, this commit adds an additional validation layer. It checks
against a Regex named IBAN2007Identifier, coming from the schema
definition.

Beware: If you use SEPA::IBANValidator for your database models too,
it may be possible that some of your data is invalid now. For
example, an IBAN containing spaces used for formatting is not allowed
anymore. But storing a formatted value in the database often is bad
practice. Make sure to store the unformatted IBAN in your database.